Extract format helpers (format_speed/format_duration/truncate_str) into format.rs
